Added news feed

+++ app/src/main/java/com/lounah/tinkoffnews/presentation/feed/list/NewsFeedViewHolders.kt
@@ -1,5 +1,5 @@
 package com.lounah.tinkoffnews.presentation.feed.list
 
-enum class NewsFeedViewHolders(val type: Long) {
+enum class NewsFeedViewHolders(val type: Int) {
     LOADING(0), FEED_ITEM(1)
 }
